PCA Inter-modal Coupling Model

Implements full volumetric IMCo coupling estimation on a single subject using PCA approach

imco_pca(
  files,
  nhoods,
  nWts,
  mask_indices,
  ref = 1,
  verbose = TRUE,
  retimg = FALSE,
  outDir = NULL,
  propMiss = NULL
)

Arguments

files
Vector of full paths to images or a list of local nifti objects

nhoods
Neighborhood matrices computed in imco

nWts
Vector of kernel weights

mask_indices
Voxels where IMCo is computed

ref
Which modality is dependent variable

verbose
TRUE for updates on computation, else FALSE

retimg
If TRUE, return list of estimated coupling maps as nifti objects

outDir
Full path to directory where maps should be written

propMiss
Maximum proportion of missing voxels in a neighborhood to tolerate, i.e., return NA if missing more than propMiss in the neighborhood of the center voxel

Value

Estimated IMCo coupling maps, either written to files and/or returned as nifti objects