Single Inter-modal Coupling model

Data for coupling estimation on a single subject and single voxel

imco_single(
  files,
  vxl,
  brainMask,
  subMask = NULL,
  ref = 1,
  neighborhoodSize = 3
)

Arguments

files
Vector of full paths to images or a list of local nifti objects

vxl
Vector of single voxel indices to get data from

brainMask
Full path to brain mask or local nifti object

subMask
NULL or local nifti object or the full path to a mask that is a subset of brain mask where coupling should be computed

ref
reference modality when type="pca" or dependent modality when type="regression"

neighborhoodSize
Full width in voxels for FWHM

Value

Data frame